# swaggerDocs.json
{
  "swagger": "2.0",
  "info": {
    "description": "FoodFactory API information",
    "version": "1.0.0",
    "title": "Food Factory"
  },
  "host": "localhost:8000",
  "basePath": "/",
...
  "paths": {
...
    "/api/ingredients/{id}": {
      "get": {
        "tags": [
          "ingredients"
        ],
        "summary": "Get ingredient details",
        "description": "Get ingredient details by matching the id",
        "operationId": "getIngredient",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ingredient id that need to be fetched"
          }
        ],
        "responses": {
          "200": {
            "description": "successful response"
          },
          "401": {
            "description": "User not logged in / User does not have admin role"
          }
        }
      },
      "put": {
        "tags": [
          "ingredients"
        ],
        "summary": "Update ingredient",
        "description": "Update ingredient for params id - only logged in admin role can update this",
        "operationId": "updateIngredient",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ingredient id that need to be updates"
          },
          {
            "name": "reqBody",
            "in": "body",
            "description": "Input ingredient details",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "availableQuantity": {
                  "type": "number"
                },
                "thresholdQuantity": {
                  "type": "number"
                },
                "price": {
                  "type": "number"
                },
                "vendorName": {
                  "type": "string"
                },
                "vendorEmail": {
                  "type": "string"
                }
              },
              "required": [
                "name",
                "availableQuantity",
                "thresholdQuantity",
                "price",
                "vendorName",
                "vendorEmail"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid input value"
          },
          "401": {
            "description": "User not logged in / User does not have admin role"
          }
        }
      },
      "delete": {
        "tags": [
          "ingredients"
        ],
        "summary": "Delete ingredient",
        "description": "Delete ingredient using params id - only logged in admin role can update this",
        "operationId": "deleteIngredient",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ingredient id that need to be deleted"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid input value"
          },
          "401": {
            "description": "User not logged in / User does not have admin role"
          }
        }
      }
    },
...
    "/api/ingredients/select/allIngredients": {
      "get": {
        "tags": [
          "ingredients"
        ],
        "summary": "Get all ingredient details",
        "description": "Get all ingredient details for admin login only",
        "operationId": "getAllIngredients",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful response"
          },
          "401": {
            "description": "User not logged in / User does not have admin role"
          }
        }
      }
    },
...
    "/api/orders/{orderId}": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Get selective order",
        "description": "Get order by id by the logged in user - admin has access to view orders by any user",
        "operationId": "getSelectiveOrder",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "description": "order id that need to be fetched"
          }
        ],
        "responses": {
          "200": {
            "description": "successful response"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "User not logged in"
          }
        }
      },
      "put": {
        "tags": [
          "orders"
        ],
        "summary": "Update order",
        "description": "Update order for selected id - can update order only of pending status",
        "operationId": "updateSelectiveOrder",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "description": "order id that needs to be updated"
          },
          {
            "name": "reqBody",
            "in": "body",
            "description": "Input for order details",
            "schema": {
              "type": "object",
              "properties": {
                "dateOfDelivery": {
                  "type": "date",
                  "example": "2020-09-27",
                  "description": "min date should be greater than order placed date"
                },
                "modeOfTransport": {
                  "type": "string"
                },
                "foodid": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": [
                "dateOfDelivery",
                "foodid"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid input value"
          },
          "401": {
            "description": "User not logged in / User does not have admin role"
          }
        }
      },
      "delete": {
        "tags": [
          "orders"
        ],
        "summary": "Delete order",
        "description": "Delete order using id - only logged in admin role can update this",
        "operationId": "deleteOrder",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "description": "Order id that needs to be deleted"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid input value"
          },
          "401": {
            "description": "User not logged in / User does not have admin role"
          }
        }
      }
    },
...
    "/api/orders/userOrders/{userId}": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Get orders by user",
        "description": "Get all the orders of a user - admin has access to view orders by any user",
        "operationId": "getUserOrders",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "description": "user id who's orders need to be fetched"
          }
        ],
        "responses": {
          "200": {
            "description": "successful response"
          },
          "400": {
            "description": "Invalid input data / invalid user id"
          },
          "401": {
            "description": "User not logged in / do not have admin role"
          }
        }
      }
    },
    "/api/orders/anyuser/{orderId}": {
      "put": {
        "tags": [
          "orders"
        ],
        "summary": "update orders by an user",
        "description": "Update order status of any user - can update order of any user only by admin",
        "operationId": "updateOrder",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "description": "order id that needs to be canceled"
          },          {
            "name": "reqBody",
            "in": "body",
            "description": "Input order status and mode of transport",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "enum" : [
                    "Pending",
                    "Shipped",
                    "Canceled",
                    "Delivered"
                  ]
                },
                "modeOfTransport": {
                  "type": "string"
                }
              },
              "required": [
                "status",
                "modeOfTransport"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid input value / cannot cancel delivered order"
          },
          "401": {
            "description": "User not logged in / User does not have admin role"
          }
        }
      }
    },
...
  },
...
}